Fallback to using defaultDomainName instead of customerId for looking up ReplaceMap.
When editing a tenant and adding tenant specific variables, the value saved is not the customerId, but rather the defaultDomainName.
This results in Get-CIPPTextReplacement not being able to find it, therefore not using it.

Ideally we should only have one true lookup identifier, but we can't just change it as that would cause issues.

Resolves https://github.com/KelvinTegelaar/CIPP/issues/4692